+++ effect_subject/tests/tests/test_followup.py
@@ -1,16 +1,17 @@
 from copy import deepcopy
 
-from django.test import TestCase
+from django.test import TestCase, tag
 from edc_constants.constants import DEAD, NO, NOT_APPLICABLE, OTHER, YES
 from model_bakery import baker
 
 from effect_screening.tests.effect_test_case_mixin import EffectTestCaseMixin
-from effect_subject.choices import INFO_SOURCES
+from effect_subject.choices import INFO_SOURCES, PATIENT_STATUSES
 from effect_subject.constants import IN_PERSON, PATIENT, TELEPHONE
 from effect_subject.forms import FollowupForm
 from effect_subject.forms.followup_form import FollowupFormValidator
 
 
+@tag("fu")
 class TestFollowup(EffectTestCaseMixin, TestCase):
     def setUp(self):
         super().setUp()
@@ -23,6 +24,7 @@ class TestFollowup(EffectTestCaseMixin, TestCase):
         form.is_valid()
 
 
+@tag("fu")
 class TestFollowupFormValidation(EffectTestCaseMixin, TestCase):
 
     form_validator_default_form_cls = FollowupFormValidator
@@ -154,22 +156,39 @@ class TestFollowupFormValidation(EffectTestCaseMixin, TestCase):
                         "info_source": info_src,
                         "info_source_other": "xxx" if info_src == OTHER else "",
                         "survival_status": DEAD,
-                        "adherence_counselling": NO,
+                        "adherence_counselling": NOT_APPLICABLE,
                     }
                 )
                 form_validator = self.validate_form_validator(cleaned_data)
                 self.assertDictEqual({}, form_validator._errors)
 
-    def test_adherence_counselling_invalid_if_deceased(self):
+    def test_adherence_counselling_na_if_deceased(self):
         cleaned_data = self.get_valid_patient_telephone_assessment_data()
         cleaned_data.update({"info_source": "next_of_kin", "survival_status": DEAD})
         form_validator = self.validate_form_validator(cleaned_data)
-        self.assertIn("adherence_counselling", form_validator._errors)
-        self.assertIn(
-            (
-                "Invalid: Adherence counselling not expected if "
-                "survival status is 'Deceased'"
-            ),
-            str(form_validator._errors.get("adherence_counselling")),
+        self.assertFieldFormValidationErrorRaised(
+            form_validator=form_validator,
+            field="adherence_counselling",
+            expected_msg="Invalid: Expected 'Not applicable' if survival status is 'Deceased'",
         )
-        self.assertEqual(len(form_validator._errors), 1, form_validator._errors)
+
+    def test_adherence_counselling_applicable_if_not_deceased(self):
+        survival_statuses = [ss[0] for ss in PATIENT_STATUSES if ss[0] != DEAD]
+        for survival_status in survival_statuses:
+            with self.subTest(survival_status=survival_status):
+                cleaned_data = deepcopy(
+                    self.get_valid_patient_telephone_assessment_data()
+                )
+                cleaned_data.update(
+                    {
+                        "info_source": "next_of_kin",
+                        "survival_status": survival_status,
+                        "adherence_counselling": NOT_APPLICABLE,
+                    }
+                )
+                form_validator = self.validate_form_validator(cleaned_data)
+                self.assertFieldFormValidationErrorRaised(
+                    form_validator=form_validator,
+                    field="adherence_counselling",
+                    expected_msg="This field is applicable.",
+                )