+++ CHANGES.rst
@@ -1,5 +1,17 @@
 Changes
 =======
 
-initial
--------
+[unreleased]
+------------
+- changes to Follow-up (formerly Clinical Assessment):
+    - rename/standardise 'Clinical Assessment' to be 'Follow-up' throughout
+    - add survival status help text, for 'Deceased' and 'Alive, but unwell' choices
+    - modify adherence_counselling to only be applicable if not 'Deceased'
+
+0.1.1
+-----
+- add migrations
+
+0.1.0
+-----
+- initial release